+++ changes.md
@@ -19,6 +19,7 @@
 13. Let ignore color "nocol" or palette "nopal" stored in the TCanvas
 14. Better support of TPaveText - lines, boxes and arbitrary text positions 
 15. Implement option "A" for histograms painters - fully disables axes drawing
+16. Support drawing of TPolyMarker
 
 
 ## Changes in 5.2.0